Implement Framework and Runner
Implement the framework and the runner for the program. Modify the
"light" class to better fit the solution. Fix the many errors in the
LightPos_API file. The API needs to be tested and the fitness functions
need to be implemented. The included Excel files provide the necessary
algorithms for the fitness functions. See code for which algorithm goes
where.